Add explicit integer-width casts to silence FPC overflow/mixing hints
Add explicit width casts at sites where FPC's "Converting the operands to
Int64 before doing the add/subtract/multiply" and "Mixing signed expressions
and longwords gives a 64bit result" hints fire because a narrow (32-bit)
sub-operation feeds a wider expression. The casts make the intended width
explicit and silence the hints on both 32-bit and 64-bit; the computed values
are unchanged on every word size.

- ClpMlDsaCore.pas (NTT): Int64(LZeta) * (Int64(LA0) - Int64(LA1)) so the
  coefficient subtract widens before the multiply.
- ClpMlKemCore.pas (compress): (Int64(LC shl n) + (MlKemQ shr 8)) so the inner
  add runs in Int64; (decompress) cast the leading shift operand to UInt32 so
  the bit-packing OR no longer mixes a signed Int32 with a longword.
- ClpPoly1305.pas: (Int64(Int32(H4 shr 26)) - 1) * 5 so the subtract is Int64.
- ClpDateTimeUtilities.pas: Int64(LOffsetHours) * 60 + LOffsetMinutes for the
  UTC offset (IncMinute already takes Int64).

Deliberately left untouched: TMod.Inverse32, TX448Field.Sub, and the ML-DSA
eta=2 mod-5 reduction. Their hints are 32-bit-only false positives over
intentional fixed-width (UInt32) modular arithmetic; FPC flags any
UInt32*UInt32 / UInt32+UInt32 on a 32-bit target regardless of casts, and
widening to Int64 would be wrong for Inverse32 (it relies on mod-2^32
wraparound). These are left as the original code rather than suppressed.

+++ CryptoLib/src/Crypto/Signers/MlDsa/ClpMlDsaCore.pas
@@ -306,7 +306,7 @@ begin
         LA0 := ACoeffs[LJ];
         LA1 := ACoeffs[LJ + LLen];
         ACoeffs[LJ] := LA0 + LA1;
-        ACoeffs[LJ + LLen] := TMlDsaReduce.MontgomeryReduce(Int64(LZeta) * Int64(LA0 - LA1));
+        ACoeffs[LJ + LLen] := TMlDsaReduce.MontgomeryReduce(Int64(LZeta) * (Int64(LA0) - Int64(LA1)));
         System.Inc(LJ);
       end;
       LStart := LJ + LLen;
